Create charmverse roles from discord roles
