+++ .eslintrc.json
@@ -47,7 +47,7 @@
     "max-len": [
       "error",
       {
-        "code": 120,
+        "code": 150,
         "ignoreComments": true,
         "ignoreStrings": true,
         "ignoreTemplateLiterals": true
